refactor: restructure Anilist provider into modular architecture
Split the 781-line anilist/mod.rs into a clean modular structure:
- models.rs: GraphQL queries, types, and shared functions (354 lines)
- people.rs: NonMediaAnilistService for staff/studio operations (333 lines)
- anime.rs: AnilistAnimeService implementation (55 lines)
- manga.rs: AnilistMangaService implementation (55 lines)
- base.rs: Core AnilistService struct (19 lines)
- mod.rs: Module exports (13 lines)

This follows the established pattern from TMDB and Hardcover refactoring,
improving code organization and maintainability while preserving all
functionality. All files compile successfully with zero clippy warnings.
